Add files via upload
We created a model via py-GM that models each player’s skill level in a dataset and can use that variable to construct the likelihood that they will beat the other player. We had two choices for judging the accuracy of our model: mean squared error between predicted probability and pairwise win rates or the percentage of the time that the player with the higher modeled skill level wins. We chose to use the latter method because of its binary nature. After this, we wanted to find the best set of parameters to optimize our model’s performance according to the previously decided upon metric. The optimal parameters were using 5 different discrete levels and having skill level scale to win probability at a rate of 0.7. Using our performance metric, this gave us a validation accuracy of 61.897%. We finally wanted to compare our graphical model to simple, non-graphical models to see how it would perform in relation to these. We made three simple models, two based on winrate and one based on total games played. Our graphical model overperformed all of these non-graphical models, overperforming 2⁄3 of them by a very significant margin.
Based on our findings, we can say with some confidence that our graphical method for modeling the skill levels of Starcraft players and using those skill levels to predict outcomes is superior to the more intuitive methods like using basic win rates or total number of games played. Our findings show that graphical models can be very useful for applications like this one.